requirements.txt
transformers
markdown-it-py
mdit_plain
openai
flask
tqdm
termcolor